refactor: moved getChannelDisplayName to a more secure place
